Started the page that displays events; lots of tech missing
This should be the start of the layout for the page that displays events. A lot of tech is missing (the APIs, the checkbox, etc, but in particular I want the page to be able to pull the event ID from the URL (like get2gather.com/events/1648264739) and update the page based on that. This should at least be a start.