Going back to old version to try and fix delete issue
There was an issue before where trying to delete an entry from the admin or invitee list would delete everything after it; I tried something different to fix it, but that broke in its own way. So I went back to this since I thought the issue was with filter() misbehaving, so I tried doing it manually.